Create materialized view in migration exclusively
